database/requirements.txt
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlite3==3.41.2